Added an example on how to use a custom retry logic for the streamer
